Vince's WGA speech in early 2025 takes a knife to the box.
JOSEPH CAMPBELL AGE 82, SUMMER 1986, SKYWALKER RANCH CALIFORNIA. LESSONS TO HOLLYWOOD FILM STAR WARS FANS

the environment was no longer spiritually responsive to the hero. The hero is today running up against a hard world that is in no way responsive to his spiritual need.
MOYERS: A windmill.
CAMPBELL: Yes, but Quixote saved the adventure for himself by inventing a magician who had just transformed the giants he had gone forth to encounter into windmills. You can do that, too, if you have a poetic imagination. Earlier, though, it was not a mechanistic world in which the hero moved but a world alive and responsive to his spiritual readiness. Now it has become to such an extent a sheerly mechanistic world, as interpreted through our physical sciences, Marxist sociology, and behavioristic psychology, that we’re nothing but a predictable pattern of wires responding to stimuli. This nineteenth-century interpretation has squeezed the freedom of the human will out of modern life.
MOYERS: In the political sense, is there a danger that these myths of heroes teach us to look at the deeds of others as if we were in an amphitheater or coliseum or a movie, watching others perform great deeds while consoling ourselves to impotence?
CAMPBELL: I think this is something that has overtaken us only recently in this culture.