Can Walter White's grave site in ABQ pull a vacuum of the Mr. Bad Ass audience sucking?
Just as he does with the “vacuum”, McLuhan takes “abcedmindedness” in two different ways, one seemingly only negative, reflecting literal (ABC…) thought, the other a difficult positive, enabling complex insight into “the gestures of being itself”.  But where he treats the “vacuum” first in negative (empty) fashion, then in positive (replete), here the order is reversed to treat “abcedmindedness” first in its positive aspect (“to be abced-minded is to be part of the dream of history that is Finnegans Wake“), then in its negative one (“just alphabetically controlled”). As with the “vacuum”, the great question is, of course, the relation between these two senses of the term.