Year 1923 Finn's Hotel Evacuation, Wake Funerals
Why did none of the clergy of the world, of any religion...

Ring the bell for a world-wide ritual of Pandemic wake? All those who could not travel to the Funeral Wakes because of shutdowns and isolation to prevent the virus spreading?

Why has the clergy not organized and become famous for holding the largest funeral rituals in human history on a planet with over 8 billion people alive? The biggest Wake Jazz Funeral Irish Finnegan party ever?

Why are we not celebrating the doctors and nurses and science that saved lives, a Thank You party and Remembrance of those lost?