Did the authors intend August 15, 1977, around 2:16 UTC reference to cut-off 78 seconds?
Vince Gilligan in interviews says what resonates with the society and individuals is how he has learned from previous series.

Vince Gilligan has demonstrated Jedi mastery of knowing what he is making.

"And to what degree is the art something that they had simply learned to do that way? When a spider makes a beautiful web, the beauty comes out of the spider’s nature. It’s instinctive beauty. How much of the beauty of our own lives is about the beauty of being alive? How much of it is conscious and intentional? That is a big question." - Gaucho Skywalker Hangout