When the Founding Fathers, Free Mason, library say the creative word of Levant John One One could come from any corner... that includes lesbians in ABQ
John 1;1 Bible verse, *from her*, she can write the Bible of any storybook or screenplay.

" So what I suddenly realized when I recognized that in the Great Seal of the United States there were two of these symbolic triangles interlocked was that we now had thirteen points, for our thirteen original states, and that there were now, furthermore, no less than six apexes, one above, one below, and four (so to say) to the four quarters. The sense of this, it seemed to me, might be that from above or below, or from any point of the compass, the creative Word might be heard, which is the great thesis of democracy. Democracy assumes that anybody from any quarter can speak, and speak truth"