Deeper deviance
Homosexuality was removed from the DSM (Diagnostic and Statistical Manual of Mental Disorders) by the American Psychiatric Association (APA) in 1973, following advocacy from LGBTQ+ activists, officially removing it as a mental illness because it was recognized as a normal variation of human sexuality, not a disorder. While listed as a "sociopathic personality disturbance" (DSM-I, 1952) and "sexual deviation" (DSM-II, 1968), the APA's 1973 vote delisted it, though a category for "Ego-Dystonic Homosexuality" (distress about one's orientation) appeared in DSM-III (1980) and DSM-III-R (1987) before being removed entirely, cementing homosexuality's status as a non-pathological orientation.