Sunday Interdependent Christmas Sermon
https://youtu.be/1jeyIAH3bUI?t=294


And We must either learn to live
together as
brothers or we All going to perish
together as
fools!

5:03
Yes, as **Nations and individuals We** are
interdependent. I've mentioned to you
before of our visit to India some years
ago it was a marvelous
experience but I say to you this morning
that thou were those depressing
moments for how can one avoid being
depressed when he sees with his own eyes
evidences of millions of people going to
bed hungry at night.     
yes sir yes sir