People talk eloquently about peace, President - what is the problem?
https://youtu.be/1jeyIAH3bUI?t=692


things that all of the great military
Geniuses of the world have talked about.
Peace the conquerors of old who came
killing in Pursuit Of Peace.
Alexander Julius Caesar charag and
Napoleon what can in Seeking a
peaceful world
order. 

11:30    
And do you know if you will read
Mein Kampf close enough Hitler
contended that everything that he did in
Germany was for
peace and the leaders of the world today
talk eloquently about
peace every time we drop our bombs in
North Viet viam! President Johnson is
talking eloquently about peace make it
it what is the
problem they are talking about
peace as a distant goal as an end we
seek but one day we must come to
see that peace is not merely a distant
goal that we seek, but it is a means by
which we arrive at that
goal. We must pursue peaceful
ends through peaceful
means! All of this is saying that in the
final
analysis means and ends must
cohere
because the end is pre-existent in the