Subtle as a New Mexico Uranium baloon
The Founding Fathers were Poets and Nobody Knows it!    

The Great Seal of the United States of America is a Battle Against Apple Plucking and Authoring of Sin, gravity of 1132.

https://www.youtube.com/watch?v=X_AfMHCVoxA       
2,990,649 views  November 26, 2021     

Carol is here to Pluck Apples and Pick Cherries   

"I hear the horses' thunder (Page 3, Finn's Hotel year 1923)     
Down in the valley below  (New Mexico)    
I'm waiting for the angels of Avalon   (omphalos)     
Waiting for the eastern glow.      
The apple of the valley hold the seeds of happiness"