Katy Perry was about her romance story of falling fro Marital Unity
"traditional Hindu wedding in India in October 2010, held near the Ranthambhore National Park at the Aman-i-Khas luxury resort, featuring elephants, camels, Indian dancers, and henna, culminating in a spiritual ceremony with Vedic chants. The six-day celebration blended Western and Indian customs, with Perry wearing a traditional red sari and Brand in Indian attire for pre-wedding parties"

==============

Legends of a "Monster State"     
In a New Mexico Valley, Twilight Zone    

.

(White House Comms Staffer) BILL MOYERS: That’s because we don’t understand the two levels of marriage.

(Sarah Lawrence Educator of Women on Men's John_sons 1:1 Word Rule of rewriting those claiming male authorships, pointing to Jean Erdman Coaching Ladies with Six Insides) JOSEPH CAMPBELL: You don’t make a commitment.

MOYERS: We presume to—we make a commitment for better or for worse.

CAMPBELL: That’s the remnant of a ritual.

MOYERS: And the ritual has lost its force. The ritual that once conveyed an inner reality is now merely form. And that’s true in the rituals of society and in the personal rituals of marriage and religion.

CAMPBELL: How many people before marriage receive spiritual instruction as to what the marriage means? You can stand up in front of a judge and in ten minutes get married. The marriage ceremony in India lasts three days. That couple is glued.

MOYERS: You’re saying that marriage is not just a social arrangement, it’s a spiritual exercise.

CAMPBELL: It’s primarily a spiritual exercise, and the society is supposed to help us have the realization. Man should not be in the service of society, society should be in the service of man. When man is in the service of society, you have a monster state, and that’s what is threatening the world at this minute.