The Hindus for example, were in a Free Mason's Free Thinking library
White House communications leader man asks some questions...

BILL MOYERS: And to them it was the god of reason [non-instrumental reason thinking in the human brain/mind].

JOSEPH CAMPBELL: Yes. This is the first nation in the world that was ever established on the basis of reason instead of simply warfare. These were eighteenth-century deists, these gentlemen. Over here we read, "In God We Trust." But that is not the god of the Bible. These men did not believe in a Fall. They did not think the mind of man was cut off from God. The mind of man, cleansed of secondary and merely temporal concerns, beholds with the radiance of a cleansed mirror a reflection of the rational mind of God. Reason puts you in touch with God. Consequently, for these men, there is no special revelation anywhere, and none is needed, because the mind of man cleared of its fallibilities is sufficiently capable of the knowledge of God. All people in the world are thus capable because all people in the world are capable of reason.

All men are capable of reason. That is the fundamental principle of democracy. Because everybody's mind is capable of true knowledge, you don't have to have a special authority, or a special revelation telling you that this is the way things should be.


MOYERS: And yet these symbols come from mythology.

CAMPBELL: Yes, but they come from a certain quality of mythology. It's not the mythology of a special revelation. The Hindus, for example, don't believe in special revelation. They speak of a state in which the ears have opened to the song of the universe. Here the eye has opened to the radiance of the mind of God. And that's a fundamental deist idea. Once you reject the idea of the Fall in the Garden, man is not cut off from his source.