Deacon Blues, University Rules, Moon Cycles and the Crimson Tides
Skywalker Ranch lessons for science fiction fandoms, 1988


they say so themselves. Read the text where it is declared that “those forms which were merely mythological forms in the past are now actual and incarnate in our Savior.” The mythologies here referred to were of the dead and resurrected god: Attis, Adonis, Gilgamesh, Osiris, one after the other. The death and resurrection of the god is everywhere associated with the moon, which dies and is resurrected every month. It is for two nights, or three days dark, and we have Christ for two nights, or three days in the tomb.

No one knows what the actual date of the birth of Jesus might have been, but it has been put on what used to be the date of the winter solstice, December 25, when the nights begin to be shorter and the days longer. That is the moment of the rebirth of light. That was exactly the date of the birth of the Persian God of light, Mithra, Sol, the Sun.


_____________

And down the coast to Duke    

I’ll try another example to evoke this postmodern scene that I will discuss again in the next lecture, not using Kierkegaard, but using a little Freud, but what the hell, it’s something. Measure the distance this way. Try to measure the distance between Wuthering Heights and the interior of reading that novel, Wuthering Heights. You know, that’s a love story, I mean it’s really a love story because it’s scary and crazy and embodied like love, its nuts. You know, the guy loves the woman so much, he’ll follow her forever and drive her crazy, even into hell if she dies. You know, it’s just a… It’s a loony, scary novel, but it has a tangible feel and it has affect, and that’s what I am coding as still real. Now, let’s compare that to Love Story. [crowd laughter].

One way to measure the historical distance we have come is by measuring the distance between Wuthering Heights and Love Story. I don’t know what you thought about the movie, but the book‘s real short. [crowd laughter]. And my own feeling was that him having a lot of money didn’t hurt. It didn’t hurt a bit. There’s an immense distance that has been travelled by human subjectivity – humans as subjects – between Wuthering Heights and Love Story, and I hate to say it, there has been some distance travelled between Love Story and now, and movies like Heathers, Blade Runner, or just visiting Los Angeles. I mean, I don’t need to use examples, just visit Los Angeles. Just walk down the walk of stars at one o’clock in the morning, and ask yourself “Are these people really here, or is this central casting?” And it isn’t a funny question. And as our society develops in this telematic way, it’s not going to be funny to ask it around the Christmas dinner table about Uncle Henry, “Is that Uncle Henry, or is it just someone playing Uncle Henry? The further postmodern insight is: what difference does it make! See, that’s the thing Reagan’s handlers understood. It doesn’t make any difference! They’re buying it! Well, that’s all I have to say for now…