Who is the one ordering sin from Eden, Romans 11:32 Bible verse
White House communications experience, Mr. Bill Moyers.

BILL MOYERS: And yet these symbols come from mythology.

JOSEPH CAMPBELL: Yes, but they come from a certain quality of mythology. It's not the mythology of a special revelation. The Hindus, for example, don't believe in special revelation. They speak of a state in which the ears have opened to the song of the universe. Here the eye has opened to the radiance of the mind of God. And that's a fundamental deist idea. Once you reject the idea of the Fall in the Garden, man is not cut off from his source.

Now back to the Great Seal. When you count the number of ranges on this pyramid, you find there are thirteen. And when you come to the bottom, there is an inscription in Roman numerals. It is, of course, 1776. Then, when you add one and seven and seven and six, you get twenty-one ...