These audio files should all be Public Domain for the sake of world peace and NYC John Lennon. Constellating images of Kepler
John Lennon was shot and killed on the night of December 8, 1980, outside his home, The Dakota apartment building, at the corner of West 72nd Street and Central Park West in New York City

__________

JOSEPH  CAMPBELL: Because it has to have constellating images to pull together all these tendencies to separation, to pull them together into some intention.

BILL MOYERS: To follow some path.

CAMPBELL: I think so. The nation has to have an intention somehow to operate as a single power.

MOYERS: What did you think of the outpouring over John Lennon’s death? Was he a hero?

CAMPBELL: Oh, he definitely was a hero.

MOYERS: Explain that in the mythological sense.

CAMPBELL: In the mythological sense, he was an innovator. The Beatles brought forth an art form for which there was a readiness. Somehow, they were in perfect tune with their time. Had they turned up thirty years before, their music would have fizzled out. The public hero is sensitive to the needs of his time. The Beatles brought a new spiritual depth into popular music which started the fad, let’s call it, for meditation and Oriental music. Oriental music had been over here for years, as a curiosity, but now, after the Beatles, our young people seem to know what it’s about. We are hearing more and more of it, and it’s being used in terms of its original intention as a support for meditations. That’s what the Beatles started.

MOYERS: Sometimes it seems to me that we ought to feel pity for the hero instead of admiration. So many of them have sacrificed their own needs for others.

CAMPBELL: They all have.