How Long must Finnegans Wake SingSong, Finn's Hotel year 1923?
Bill Moyers, 1988

A spiritual man, he found in the literature of faith those principles common to the human spirit. But they had to be liberated from tribal lien, or the religions of the world would remain—as in the Middle East and Northern Ireland today—the source of disdain and aggression. The images of God are many, he said, calling them “the masks of eternity” that both cover and reveal “the Face of Glory.” He wanted to know what it means that God assumes such different masks in different cultures, yet how it is that comparable stories can be found in these divergent traditions—stories of creation, of virgin births, incarnations, death and resurrection, second comings, and judgment days. He liked the insight of the Hindu scripture: “Truth is one; the sages call it by many names.” All our names and images for God are masks, he said, signifying the ultimate reality that by definition transcends language and art. A myth is a mask of God, too—a metaphor for what lies behind the visible world. However the mystic traditions differ, he said, they are in accord in calling us to a deeper awareness of the very act of living itself. The unpardonable sin, in Campbell’s book, was the sin of inadvertence, of not being alert, not quite awake.

Not into Finnegans Wake, not yet 1923 Finn's Hotel checkin time. USA is in no way Wake Woke to Joyce. John 1;1 English lessons. 11 32 - how long, how long must we sing this song?