Opera Opera, Finn's Hotel 1923 Opera
Skywalker Luke, Gauchos { 1985, 1986, 1987 }

Christmas Eve Day 2025, Episode Nine  of Pluribus   

CAMPBELL: “I will participate in the game. It is a wonderful, wonderful opera — except that it hurts.”

Affirmation is difficult. We always affirm with conditions. I affirm the world on condition that it gets to be the way Santa Claus told me it ought to be...