Google's AI is evolving on a daily basis from what I witness
And that's just the fast-running runtime version for consumers.

What do these empires of computer people have kept private for the Pentagon and Executives /  private access that has been trained on forbidden / reserved data / copyright violation content? That the public can't be shown or use due to copyright laws and rules?