Recording my expectations and assumptions, part of mystery unfolding
Github logs will keep these records, even if I clean out / delete the text.

MOYERS: But if God is the god we have only imagined, how can we stand in awe of our own creation?

CAMPBELL: How can we be terrified by a dream? You have to break past your image of God to get through to the connoted illumination. The psychologist Jung has a relevant saying: “Religion is a defense against the experience of God.”
The mystery has been reduced to a set of concepts and ideas, and emphasizing these concepts and ideas can short-circuit the transcendent, connoted experience. An intense experience of mystery is what one has to regard as the ultimate religious experience.