Vince Gilligan / Pluribus: THAT IS BIG STUFF! WOW VIRTUE WEST GATE SIGNALING!
And Tristan says, “By my death, do you mean this pain of love?” — because that was one of the main points, that one should feel the sickness of love. There’s no possible fulfillment in this world of that identity one is experiencing. Tristan says, “If by my death, you mean this agony of love, that is my life. If by my death, you mean the punishment that we are to suffer if discovered, I accept that. And if by my death, you mean eternal punishment in the fires of hell, I accept that, too.”