Four seasons
The phone call again when the hive mind crashed I think shows this is going for global mythology. We have Elvis in the wings... Hindu / Buddhism. We were told Zosia lost someone, but not sure why she is in Africa dressed like Islam. The Peruvian woman represented native non-Levant cultures.