Egg walls invaded, sperm penetrated the walls, needle penetrated the walls of Carol's age 16 abusive parents childhood
Carol's parents could understand the lesbian love, they only loved like the hive mind loves. Generic sex love, not  Troubadour Year 1210 individual laws of love.

Carl has met her self at puberty now, she is seeing into the future of children now. Parents who are blind to the Year 1210 Troubadours ideal of Love.

Outer-space supernaturally imposed by machines flying overhead at the Zero Hour.

Power of Myth

MOYERS: When I listen to you talk about how myths connect us to our sacred places, and how landscapes connected primal human beings to the universe, I begin to think that the supernatural, at least as you understand it, is really only the natural.

CAMPBELL: The idea of the supernatural as being something over and above the natural is a killing idea. In the Middle Ages this was the idea that finally turned that world into something like a wasteland, a land where people were living inauthentic lives, never doing a thing they truly wanted to because the supernatural laws required them to live as directed by their clergy. In a wasteland, people are fulfilling purposes that are not properly theirs but have been put upon them as inescapable laws. This is a killer. The twelfth-century troubadour poetry of courtly love was a protest against this supernaturally justified violation of life’s joy in truth. So too the Tristan legend and at least one of the great versions of the legend of the Grail, that of Wolfram von Eschenbach. The spirit is really the bouquet of life. It is not something breathed into life, it comes out of life. This is one of the glorious things about the mother-goddess religions, where the world is the body of the Goddess, divine in itself, and divinity isn’t something ruling over and above a fallen nature. There was something of this spirit in the medieval cult of the Virgin, out of which all the beautiful thirteenth-century French cathedrals arose.

However, our story of the Fall in the Garden sees nature as corrupt; and that myth corrupts the whole world for us.