Leave it alone, can't see Great Seal Eye to Eye, no good Carol no bad Carol
Power of Myth, Skywalker Ranch California lessons for Science Fiction audiences 

The Creative Word of Carol's Word writing, Helen's minute 22 post-graduation thesis of Episode One at the Zero Hour Public house: Creative Words of Finn's Hotel year 1923...

"How about you? Have you got a place to stay? Why should I care?" - "It's the end of our love song"

------

"the initial sound (a Christian might say, the creative Word), out of which the whole world was precipitated, the big bang, the pouring of the transcendent energy into and expanding through the field of time. As soon as it enters the field of time, it breaks into pairs of opposites, the one becomes two. Now, when you have two, there are just three ways in which they can relate to one another: one way is of this one dominant over that; another way is of that one dominant over this; and a third way is of the two in balanced accord. It is then, finally, out of these three manners of relationship that all things within the four quarters of space derive.

There is a verse in Lao-tzu’s Tao-te Ching which states that out of the Tao, out of the transcendent, comes the One. Out of the One come Two; out of the Two come Three; and out of the Three come all things."