Create README.md
Just in Time Delivery for Year 250, 1776 to 2026 July 4

Saturday, I think it was July 4, 2025 

In the park, Every Day was Great Seal of the Untied States of America day for Creative Mythology awakening.

Get 1923 Finn's Hotel Woke!!!

https://www.youtube.com/watch?v=HjylD7esXDo

5,857,064 views  May 29, 2019  

Do you recall the lead-up to Bloomsday 2019? The Eye of the Great Seal recalls Saturday in the Phoenix Dublin Park!