Vince Gilligan takes Television and Social Media messaging seriously enough, Vince Gilligan went to NYU
"And so, I raise no objection to television's junk. The best things on television are its junk, and no one and nothing is seriously threatened by it. Besides, we do not measure a culture by its output of undisguised trivialities but by what it claims as significant. Therein is our problem, for television is at its most trivial and, therefore, most dangerous when its aspirations are high, when it presents itself as a carrier of important cultural conversations. The irony here is that this is what intellectuals and critics are constantly urging television to do. The trouble with such people is that they do not take television seriously enough." - Page 16. New York University, 1985 published. The year Vince Gilligan graduated high school.

https://old.reddit.com/r/nosurf/comments/g4ndbp/amusing_ourselves_to_death/fnzlyl7/

April 20, 2020 pandemic expanding days 

https://books.google.com/books?id=oup6iagfox8C&pg=PA16&dq=%22The+best+things+on+television+are+its+junk,+and+no+one+and+nothing+is+seriously+threatened+by+it.%22&hl=en&newbks=1&newbks_redir=1&sa=X