Think-giving off by one error
An "off-by-one error" is a common programming bug where a value is exactly one more or one less than intended, often due to improper loop conditions or confusion over zero-based indexing. 

The "fencepost problem": This classic example of an off-by-one error (a 10-foot fence needs 11 posts, not 10) could be a metaphor for situations where the expected number of components is consistently wrong because an endpoint was forgotten. 

https://en.wikipedia.org/wiki/Off-by-one_error

...

"She knew all the Stella bros before any of us. I don’t know exactly how it came up, but I guess she had shown them some of our demos, and they rocked with it and asked if we wanted to do a split. The rest is history. As for planning what songs to put on there, we wrote Weight of Too Many specifically for the split, Off By One Again on the other hand had been a song we had written before a bit prior. We just felt it matched the vibe of what we wanted for the split and complemented Weight of Too Many so well, so we threw it on there."

https://old.reddit.com/r/Emo/comments/1pec33y/the_ritonello_form_off_by_one_again_ffo_braid/
December 4, 2025