A dash of context between quote and citation of author
Some people can't see 1993 and the behavior since year 2013 

The context of shifting media of Apple pie iPhone iPad beholding "making things great" again through pretzel pie logic of import Levant 2,000 year ago conflict patterns, Babel towers.

Making exactly what great again? The Great Seal of the United States of America is Great at cherry picking.

We The People are following signals of fixated self-harm. Since year 2012 / year 2013. Marshall McLuhan, University of Toronto, next door to the USA observing... "mass man", "it is called being mass man" from high-speed electric media.


.

"Caetextia presents itself as the inability to adjust behaviour appropriately to deal with interacting variables. People with caetextia may fail to consider the context surrounding the behaviour. This can result in people with caetextia experiencing elevated levels of frustration, anger, and anxiety when faced with a situation that requires giving attention to more than one interacting variable or factor at a time. This can be attributed to the inability to unconsciously draw upon the contextual information presented in a given situation as well as evaluate the significance of change with regards to the surrounding environment."