It has become hard. ITS HARD 1982. Eminence Front! Front page README.md message before New Years 2026 of USA 250 Saturday July 4
it was not a mechanistic world in which the hero moved but a world alive and responsive to his spiritual readiness. Now it has become to such an extent a sheerly mechanistic world, as interpreted through our physical sciences, Marxist sociology, and behavioristic psychology

   https://www.youtube.com/watch?v=BfrUQA2tb6M    
31,107,319 views July 31, 2018

Saturday July 4, 2026 - Every Day became the Great Seal of the United States of America self-awareness day!