The Vince Gilligan Saul lawyer-speak MAGA USA Truth Social fandom
YouTube user who wants all other You Tube users to speak the hive mind language.

Like forcing English from King James upon Native American peoples in New Mexico....

Campbell 1987:

These missions are mystical journeys with all of the details of the typical mystical journey. First, there is disengagement from secular life. Everybody who is going to go on this expedition has to make a complete confession of all the faults of his or her recent living. And if they don’t, the magic is not going to work. Then they start on the journey. They even speak a special language, a negative language. Instead of saying yes, for example, they say no, or instead of saying, “We are going,” they say, “We are coming.” They are in another world.
Then they come to the threshold of the adventure.


We Us / All of We The People, Pluribus... even speak a special language, a negative language.