Power of Myth year 1988 book, page 243
THAT Helen Woman or THAT Pirate woman or the Vegas Elvis Suite mix and match?

---

I have had a revelation from my computer about mythology. You buy a certain software, and there is a whole set of signals that lead to the achievement of your aim. If you begin fooling around with signals that belong to another system of software, they just won’t work.

Similarly, in mythology — if you have a mythology in which the metaphor for the mystery is the father, you are going to have a different set of signals from what you would have if the metaphor for the wisdom and mystery of the world were the mother. And they are two perfectly good...

_____
__________
_________________

The third function is the sociological one — supporting and validating a certain social order. And here’s where the myths vary enormously from place to place. You can have a whole mythology for polygamy, a whole mythology for monogamy. Either one’s okay. It depends on where you are. It is this sociological function of myth that has taken over in our world—and it is out of date.

The Elvis Suite has a different social order...