Minted in year 179 gold
The phrase "E Pluribus Unum" (Out of Many, One) first appeared on U.S. coins in 1795, on the $5 gold (half eagle) coin, linked to the Great Seal, and was mandated on all coins by law in 1873; it also appeared on the back of the $1 bill starting in 1935, as part of the Great Seal design. 

.... ....

"riverrun, past Eve and Adam's, from swerve of shore to bend of bay, brings us by a commodius vicus of recirculation back to Howth Castle and Environs"
......... 

In circulation

The value of a $5 gold (half eagle) coin can range from approximately $1,150 to over $960,000, depending heavily on its specific type, year, mint mark, and condition. The value is primarily driven by its rarity and numismatic (collector) demand, far exceeding its intrinsic gold melt value