Age 44 mother? Age 9 child? India...
I think it is safe to assume this plot point isn't going to be muddled that much. Some things in this show are wide open:

1. Why 4 stars are missing of the 13 next to the eagle head on the POTUS Seal symbol?

2. Are space aliens at war? Are star systems sending competing messages? Show can go a lot of ways.

But this plot line seems pretty clear to predict / project into the future seasons of the show

She will choose to change her mind about her age  9 child being int he hive mind and choose NOT to join the hive mind and to somehow get her child out.

That's the inward journey I see.