The Mechanical LSD Acid Trip, She is your Bride in the Sex Shop for your boy
Power of Myth, 1988 book

CAMPBELL: The old-time religion belongs to another age, another people, another set of human values, another universe. By going back you throw yourself out of sync with history. Our kids lose their faith in the religions that were taught to them, and they go inside.

BILL MOYERS: Often with the help of a drug.

CAMPBELL: Yes. The mechanically induced mystical experience is what you have there. I have attended a number of psychological conferences dealing with this whole problem of the difference between the mystical experience and the psychological crack-up. The difference is that the one who cracks up is drowning in the water in which the mystic swims. You have to be prepared for this experience.