Power Of Myth, Page 62, too good to be true
Year 1961, New York City 

....

These functions then, mythology serves. For all individuals, there have been functions to this day. And it does not matter whether the stories told are true or false — they work, just the way the movie or a play works upon you. I know that many of us like to feel that the stories that we read are true stories. There are, for example, True Story magazines. But many of the stories in True Story Magazine are not true at all. And yet we enjoy them that way just as well. Now I would like to propose that mythologies serve whether they are true or false.

Dante, in his analytical work the Convivio, said that there are two ways of regarding the literal aspect of a mythological image: one is the way of the poet, and the other is the way of the theologian.

The poet sees the literal story as a beautiful fiction through which a truth is communicated allegorically.

The theologian sees the story as a fact through which a truth is communicated.

Both of them point to a superior truth through the story, but for one the story must be true — he likes True Story magazines; the theologians like True Story magazines. And artists like fiction.

My own personal definition of mythology — of religion — is religion is a popular misunderstanding of poetry. Well now the poetry works, and so does the True Story magazine.

And I have reviewed very briefly the working — the operation of these images now in organizing the society by telling you that this is the good society, and that it is good to do this, and not good to do that, and that that is the bad society, and you are doing a good thing when you kill its members.

The other function which the fictions serve is that of balancing, coordinating, and giving enriching imagery and food to the psyche.