But the mechanical devices we use to make warfare are very true stories that are not exagerated
1988 Book, again, Finnegans Wake page 628, GOTO page 3

BILL MOYERS: The Indians addressed the buffalo as “thou,” an object of reverence.

CAMPBELL: The Indians addressed all of life as a “thou”—the trees, the stones, everything. You can address anything as a “thou,” and if you do it, you can feel the change in your own psychology. The ego that sees a “thou” is not the same ego that sees an “it.” And when you go to war with people, the problem of the newspapers is to turn those people into “its.”

MOYERS: This happens in marriage, too, doesn’t it? And happens with children, too.

CAMPBELL: Sometimes the “thou” turns into an “it,” and you don’t know what the relationship is. T