Teaching Dreams, American Dreams Teaching
Bill Moyers worked at the White House in Washington D.C.   
He only died this year 2025.

"reconstruct an order of initiation that would result in spiritual revelation"

BILL MOYERS: What explains the relationship between these symbols and the Masons, and the fact that so many of these founding fathers belonged to the Masonic order? Is the Masonic order an expression somehow of mythological thinking?

CAMPBELL: Yes, I think it is. This is a scholarly attempt to reconstruct an order of initiation that would result in spiritual revelation. These founding fathers who were Masons actually studied what they could of Egyptian lore. In Egypt, the pyramid represents the primordial hillock. After the annual flood of the Nile begins to sink down, the first hillock is symbolic of the reborn world. That's what this seal represents.