To not upset the apple cart (Apple TV+ Audience, Apple //+ device addicts audience)
censorship bypass of using the "Kepler 22B alien story" to bypass the censorship in the hearts and minds of society...

That's what Campbell was pointing out about the "love potion" BEFORE the person to person attraction.

To bypass censorship, to prevent the story from being crushed.

How many stories were unpopular, like my GitHub is unpopular? That were never preserved, never had enough copies printed to make it to the year 2025?

A lot of labor an wealth and luck has to go into year 1100 stories being preserved or known at all. Drowned out by powers that be and attraction to Fox News HDTV of the time.